Add files via upload
This Part Of Application Consist Of Child Behaviour & Posture Analysis Of a missing child, which includes various posture analyses like Sitting Crouching and Various facial analysis also added to it. It also Consist Of A aditional Frontend Added and Google Voice TTS added to it, Along With Data Viusalization, which can be accessed through some simple steps to access the code, and for its detailed explanation. Please access the following git account by the URL and follow some simple steps to get it hands On:

https://github.com/shotokan2003/vedant